requirements.txt
Flask==2.3.3
python-docx==0.8.11
docx2txt==0.8
reportlab==4.0.4